added currentUser context
